Refactored database.py to db_operations.py
